Arcade-scoped stats.
