+++ config/routes.rb
@@ -30,6 +30,7 @@ Rails.application.routes.draw do
     member do
       get :confirm_destroy
       get :images
+      get :stats
 
       resources :approval_requests, only: [:new, :update, :show], controller: :approval_requests
     end
